Extract SchemaUnifier
